Implement Exception Handling in ApplicationController
- Introduced ExceptionHandler module to centralize error handling logic.
- Moved error handling from ApplicationController to the new ExceptionHandler concern.
- Created Unauthorized error class for better clarity in authorization errors.
- Updated ApplicationController to include the ExceptionHandler and removed inline error handling code.
