+++ app/controllers/application_controller.rb
@@ -1,7 +1,7 @@
 class ApplicationController < ActionController::API
+  include ExceptionHandler
+  
   before_action :authenticate_user!
-
-  class Unauthorized < StandardError; end
 
   def current_user
     @current_user ||= User.find_by(id: header_user_id)
@@ -16,20 +16,5 @@ class ApplicationController < ActionController::API
 
   def header_user_id
     request.headers["X-User-Id"]
-  end
-
-  rescue_from StandardError do |e|
-    render json: { 
-      error: e.message, 
-      backtrace: e.backtrace.first(10) 
-    }, status: :internal_server_error
-  end
-
-  rescue_from ActiveRecord::RecordNotFound do |e|
-    render json: { error: e.message }, status: :not_found
-  end
-
-  rescue_from Unauthorized do
-    render json: { error: "unauthorized" }, status: :unauthorized
   end
 end
