+++ app/errors/unauthorized.rb
@@ -1,0 +1,1 @@
+class Unauthorized < StandardError; end
